% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotFF}
\alias{plotFF}
\title{Plot Function}
\usage{
plotFF(data = NULL, data2 = NULL, gene = NULL, FPC = NULL)
}
\arguments{
\item{data}{Longitudinal gene expression data for type one patients.}

\item{data2}{Longitudinal gene expression data for type two patients.}

\item{gene}{Functional scores for the patients.}

\item{FPC}{A FPCA object returned by the FPCA function.}
}
\description{
\code{plotFF} generates the fitted lines on the data using FPCA.
}
\keyword{internal}
