% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simudata}
\alias{simudata}
\title{Function to Simulate Testing Data}
\usage{
simudata(n = 50, nf = 5)
}
\arguments{
\item{n}{Total number of patients.}

\item{nf}{The average number of follow-up visits per patient.}
}
\value{
returns a list with following objects.
\item{data}{Longitudinal gene expression data.}
\item{data.id}{Survival data with important covariates.}
}
\description{
\code{simudata} is a function that generates a simulated data.
}
\examples{
data.list = simudata()
}
\references{
LCox: A tool for selecting genes related to survival outcomes
using longitudinal gene expression data. Jiehuan Sun, Jose D. Herazo-Maya,
Jane-Ling Wang, Naftali Kaminski, and Hongyu Zhao.
}
